server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.44.1",
    "grpcio>=1.74.0",
    "grpcio-tools>=1.74.0",
    "openai>=1.102.0",
    "openai-agents>=0.2.10",
]
